// src/app/page/master/map-angka-to-romawi/map-angka-to-romawi.interface.ts
import { EventEmitter } from '@angular/core';

export interface MapAngkaToRomawi {
	kdProduk?: string;
	label?: any;
	value?: any;
	kdMapping?: any;
	kode?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}